Add client struct and use to call api methods
